fixes for compilation with clang

+++ src/libmaus/util/DigitTable.cpp
@@ -21,14 +21,14 @@
 libmaus::util::DigitTable::DigitTable()
 {
 	memset(&A[0],0,sizeof(A));
-	A['0'] = 1;
-	A['1'] = 1;
-	A['2'] = 1;
-	A['3'] = 1;
-	A['4'] = 1;
-	A['5'] = 1;
-	A['6'] = 1;
-	A['7'] = 1;
-	A['8'] = 1;
-	A['9'] = 1;
+	A[static_cast<int>('0')] = 1;
+	A[static_cast<int>('1')] = 1;
+	A[static_cast<int>('2')] = 1;
+	A[static_cast<int>('3')] = 1;
+	A[static_cast<int>('4')] = 1;
+	A[static_cast<int>('5')] = 1;
+	A[static_cast<int>('6')] = 1;
+	A[static_cast<int>('7')] = 1;
+	A[static_cast<int>('8')] = 1;
+	A[static_cast<int>('9')] = 1;
 }
